+++ controller/konnect/constraints/constraints.go
@@ -23,6 +23,7 @@ type SupportedKonnectEntityType interface {
 		configurationv1alpha1.KongCredentialAPIKey |
 		configurationv1alpha1.KongUpstream |
 		configurationv1alpha1.KongCACertificate |
+		configurationv1alpha1.KongCertificate |
 		configurationv1alpha1.KongTarget |
 		configurationv1alpha1.KongVault
 	// TODO: add other types